requirements.txt
flask
pillow
transformers
tensorflow
tf-keras
torch
accelerate